Work on adding an AddressSet model.
models.py: created the AddressSet model, added a required foreign key for addresses to be members of a set.
all_addresses.html: added the description of the set to the table. (Is there a template tag to dump all the fields of a model without writing a bunch of <td> tags?)
tests.py: Refactored with a base class providing basic testing DB setup methods, accommodated the new foreign-key constraint for Address.