Major refactor of google_usps_validator
1. Clean up the decision tree for google_geocode(). Use dedicated
methods to set the status, messages, and address for any given outcome.
Use method calls to these in the decision tree.
2. Fix the ROOFTOP bug. If there are high-quality results (ROOFTOP),
then filter the result set to include only those.
3. Use object properties to hold the important information about the
object (address, status, result, user_input) because there are a lot of
methods needing to access these.
4. Clean up the possible statuses. Possible statuses after the Google
stagea are now MAPPED and MAPPED_PARTIAL. Fail statuses are UNSUBMITTED
and FAILED. Add room to the object's status field in anticipation of
what's to come.
5. Set the user_input on a GoogleUspsValidator as an object variable,
then call validate() instead of validate(address).