Set up static files and a base CSS file
Not included in this commit: a surprising amount of futzing about with
settings.py and nginx settings.
nginx needs to have /static/ pointed to the project's static directory,
and settings.py needs to have

STATIC_URL =  '/static/'
STATIC_ROOT = the same directory as specified in nginx.

And then you need to run manage.py collectstatic, and static files
should work.