Fancy colors!
Change color of the headline on the address_detail page according to
validation status.

Use the standard representation of the address on all_addresses, which
should be the validated version if status is MATCHED.

Add a colored bar next to the row in all_addresses to show the
validation status!

Remove pk column from all_addresses

Improve formatting for address_detail: move status up to a subhead