A large number of updates to the client side
Added a bunch of hidden fields with more detailed address data to use for nicer formatting in the table.
Changed the "validate" column to the more generic "action" column so that you can do all necessary actions from the main page.
Added class names to the data columns for easier referencing with selectors
Added proper dataTables-bootstrap integration.
updateStatusClasses now removes any applicable classes before adding new ones, to avoid multiple classes triggering wrong behaviors on the same cell.
updateStatusClasses now adds helper classes to the "action" column to direct the onClick handlers.
Factored out initialization settings for the dataTable into separate variables for readability.
Added JS functions to update the contents of the address cell and the action cell, depending on the status of the address
Added customization options on the export buttons so that a useful subset of the data columns could be exported for use.
Adjusted column widths so that the address column wouldn't break mid-line for most addresses
Customized Bootstrap's default dom settings for the dataTable
Increased the minimum list presentation size to something reasonably large